+++ R/use_python.R
@@ -18,6 +18,9 @@ use_python <- function(python, required = FALSE) {
   if (required && !file_test("-f", python) && !file_test("-d", python))
     stop("Specified version of python '", python, "' does not exist.")
 
+  if (required)
+    .globals$required_python_version <- python
+  
   .globals$use_python_versions <- unique(c(.globals$use_python_versions, python))
 }